show last sample value in plot when the cursor is higher that the last channel time stamp
