+++ asammdf/blocks/mdf_common.py
@@ -151,42 +151,3 @@ class MDF_Common:
 
         return gp_nr, ch_nr
 
-    def whereis(self, channel, source=None):
-        """get ocurrences of channel name in the file
-
-        Parameters
-        ----------
-        channel : str
-            channel name string
-        source (None) : str
-            filter the channel occurences using the source name
-
-            .. versionadded:: 6.0.0
-
-        Returns
-        -------
-        ocurrences : tuple
-
-
-        Examples
-        --------
-        >>> mdf = MDF(file_name)
-        >>> mdf.whereis('VehicleSpeed') # "VehicleSpeed" exists in the file
-        ((1, 2), (2, 4))
-        >>> mdf.whereis('VehicleSPD') # "VehicleSPD" doesn't exist in the file
-        ()
-
-        """
-
-        if channel in self:
-            if source:
-                try:
-                    entry = self._validate_channel_selection(channel, source=source)
-                except:
-                    return tuple()
-                else:
-                    return (entry, )
-            else:
-                return tuple(self.channels_db[channel])
-        else:
-            return tuple()
