+++ asammdf/mdf.py
@@ -8,6 +8,7 @@ from datetime import datetime, timezone
 from functools import reduce
 import logging
 from pathlib import Path
+import re
 from shutil import copy
 from struct import unpack
 import xml.etree.ElementTree as ET
@@ -4018,6 +4019,46 @@ class MDF:
         if self._callback:
             out._callback = out._mdf._callback = self._callback
         return out
+
+    def whereis(self, channel, source=None):
+        """get ocurrences of channel name in the file
+
+        Parameters
+        ----------
+        channel : str
+            channel name string
+        source (None) : str
+            filter the channel occurences using the source name
+
+            .. versionadded:: 6.0.0
+
+        Returns
+        -------
+        ocurrences : tuple
+
+
+        Examples
+        --------
+        >>> mdf = MDF(file_name)
+        >>> mdf.whereis('VehicleSpeed') # "VehicleSpeed" exists in the file
+        ((1, 2), (2, 4))
+        >>> mdf.whereis('VehicleSPD') # "VehicleSPD" doesn't exist in the file
+        ()
+
+        """
+
+        if channel in self:
+            if source:
+                try:
+                    entry = self._validate_channel_selection(channel, source=source)
+                except:
+                    return tuple()
+                else:
+                    return (entry, )
+            else:
+                return tuple(self.channels_db[channel])
+        else:
+            return tuple()
 
 
 if __name__ == "__main__":
